+++ nextstep-mvc/src/main/java/nextstep/mvc/handler/AbstractHandlerAdapter.java
@@ -1,5 +1,7 @@
 package nextstep.mvc.handler;
 
+import java.util.Optional;
+
 public abstract class AbstractHandlerAdapter implements HandlerAdapter {
     private final Class<?> clazz;
 
@@ -9,6 +11,8 @@ public abstract class AbstractHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean apply(Object handler) {
+        Optional.ofNullable(handler)
+                .orElseThrow(() -> new IllegalArgumentException("지원하지 않는 핸들러 입니다."));
         return clazz.isInstance(handler);
     }
 }
