+++ nextstep-mvc/src/main/java/nextstep/mvc/handler/HandlerAdapter.java
@@ -1,6 +1,6 @@
 package nextstep.mvc.handler;
 
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
