+++ nextstep-mvc/src/main/java/nextstep/mvc/handler/LegacyHandler.java
@@ -1,10 +1,10 @@
 package nextstep.mvc.handler;
 
 import nextstep.mvc.asis.Controller;
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.mvc.tobe.RedirectView;
-import nextstep.mvc.tobe.View;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.mvc.view.RedirectView;
+import nextstep.mvc.view.View;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
