+++ nextstep-mvc/src/main/java/nextstep/mvc/view/JspView.java
@@ -1,21 +1,21 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.view;
 
-import org.apache.commons.lang3.StringUtils;
-
+import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Map;
 
-public class RedirectView implements View {
-    private static final String PREFIX_REDIRECTION = "redirect:";
+public class JspView implements View {
     private final String path;
 
-    public RedirectView(String path) {
+    public JspView(String path) {
         this.path = path;
     }
 
     @Override
     public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        response.sendRedirect(StringUtils.stripStart(path, PREFIX_REDIRECTION));
+        RequestDispatcher rd = request.getRequestDispatcher(path);
+        model.forEach(request::setAttribute);
+        rd.forward(request, response);
     }
 }
