+++ nextstep-mvc/src/main/java/nextstep/mvc/view/ModelAndView.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.view;
 
 import java.util.Collections;
 import java.util.HashMap;
